third_party/blink/web_tests/external/wpt/fullscreen/api/document-fullscreen-keyboard-lock-shadowdom.tentative.window-expected.txt

This is a testharness.js-based test.
[FAIL] Requesting and exiting fullscreen with different keyboard lock options in Shadow DOM should work as expected
  assert_equals: fullscreen should deactivate expected null but got Element node <div></div>
Harness: the test ran to completion.
